Prune unneeded runtime dependencies (#192)
Dependency pruning, library removal, install size reduction and speedup

* Stop pretending there's documentation

Remove docs scripts, documentation.yml and the documentation dependency.

* Update node-factorio-api

The 0.3.5 release of node-factorio-api removed the dependency to
documentation, which itself pulls in another 500 dependencies.  Update
to 0.3.5 to avoid having this pulled in.

* Remove snyk

It's not supposed to be installed as a dependency in the first place,
and this seems like the wrong solution to there being too many stinking
dependencies to keep track the security of.

* Reimplement hashFile locally

Replace the 10 MB of dependencies that is hash-files with a homebrew
Promise based implementation of hashFile that uses crypto and fs.  This
somehow also ends up being the same amount of code as the one using
hash-files.

* Move html5-validator to devDependencies

html5-validator is only used for testing the master server web interface
and does not need to be included on production installs.

* Use --only=production on npm install

Add --only=production flag to Dockerfile and the README entries
describing how to install nodes.  This will avoid pulling in a lot of
dependencies that are not needed for running master/slaves.

* Remove pointless micro-optimization

This saved 30ms of the startup on my machine.  Even if it saved a second
it's still pretty pointless as we're not depending on the master server
startup being super fast.

* Remove undescore

A 50 kb library is not needed to make a function run once, especially
not when it doesn't work.  _.once does not modify the function passed to
it to only run once it returns a new function that only runs once.
Since the result of once was discarded, it had no effect.

Remove underscore and implement instanceManagement running once properly
with 3 lines of code.

* Remove bcryptjs

Remove unused dependency bcryptjs.

* Remove app-root-path

Resolve the relative path using path.relative instead of some hacky
string substitutions.

* Remove right-pad

Remove unused dependency right-pad.