Fix 'Got no "ok"' log spam
In 1bdf23d I incorrectly assumed that the response sent wouldn't matter
as it hadn't sent a response in ages.  That was not the case and the
client code checks to see if the response is "ok" while .sendStatus(200)
sends a response of "OK".  Fix log spam by doing a case insensitive
comparison of the response.